relocate components
