+++ src/app/(admin)/admin/users/create/page.jsx
@@ -1,5 +1,5 @@
 import { cookies } from "next/headers";
-import CreateUserForm from "@/app/components/templates/UserCreateCMS";
+import CreateUserForm from "@/app/components/forms/CreateUserFormCMS";
 
 export default async function CreateUserPage() {
   const cookieStore = await cookies();
