+++ src/app/components/forms/EditUserFormCMS.tsx
@@ -4,12 +4,12 @@ import Link from "next/link";
 import Image from "next/image";
 import { useRouter } from "next/navigation";
 import { setSessionToken, trpc } from "@/trpc/client";
-import AppButton from "@/app/components/elements/AppButton";
-import TitleRevealCMS from "@/app/components/elements/TitleRevealCMS";
-import InputCMS from "@/app/components/elements/InputCMS";
-import SelectCMS from "@/app/components/elements/SelectCMS";
-import StatusLabelCMS from "@/app/components/elements/StatusLabelCMS";
-import TextAreaCMS from "@/app/components/elements/TextAreaCMS";
+import AppButton from "@/app/components/buttons/AppButton";
+import TitleRevealCMS from "@/app/components/titles/TitleRevealCMS";
+import InputCMS from "@/app/components/fields/InputCMS";
+import SelectCMS from "@/app/components/fields/SelectCMS";
+import StatusLabelCMS from "@/app/components/labels/StatusLabelCMS";
+import TextAreaCMS from "@/app/components/fields/TextAreaCMS";
 import { Switch } from "@/components/ui/switch";
 import { toast } from "sonner";
 import {
@@ -24,17 +24,21 @@ import {
   Save,
   ChevronRight,
 } from "lucide-react";
-import UploadImageCMS from "../elements/UploadAvatarUserCMS";
-import UploadAvatarUserCMS from "../elements/UploadAvatarUserCMS";
-import AppBreadcrumb from "../elements/AppBreadcrumb";
-import AppBreadcrumbItem from "../elements/AppBreadcrumbItem";
+import UploadAvatarUserCMS from "../fields/UploadAvatarUserCMS";
+import AppBreadcrumb from "../navigations/AppBreadcrumb";
+import AppBreadcrumbItem from "../navigations/AppBreadcrumbItem";
 
-interface CreateUserFormProps {
+interface EditUserFormProps {
   sessionToken: string;
+  userId: string;
 }
 
-export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
-  const createUser = trpc.create.user.useMutation();
+export default function EditUserForm({
+  sessionToken,
+  userId,
+}: EditUserFormProps) {
+  const utils = trpc.useUtils();
+  const editUser = trpc.update.user.useMutation();
   const router = useRouter();
 
   // --- Set session token to client
@@ -44,7 +48,12 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
     }
   }, [sessionToken]);
 
-  // --- Return data from tRPC
+  // --- Return initial data
+  const {
+    data: initialData,
+    isLoading: isLoadingInitial,
+    isError: isErrorInitial,
+  } = trpc.read.user.useQuery({ id: userId }, { enabled: !!sessionToken });
   const {
     data: rolesData,
     isLoading: isLoadingRoles,
@@ -77,17 +86,41 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
     industry: string | number;
     entrepreneurStage: string | number;
   }>({
-    fullName: "",
-    email: "",
-    avatar: "",
-    roleId: "",
-    status: "ACTIVE",
-    dateOfBirth: "",
-    learningGoal: "",
-    businessName: "",
-    industry: "",
-    entrepreneurStage: "",
+    fullName: initialData?.user.full_name || "",
+    email: initialData?.user.email || "",
+    avatar: initialData?.user.avatar || "",
+    roleId: initialData?.user.role.id ?? "",
+    status: initialData?.user.status || "ACTIVE",
+    dateOfBirth: initialData?.user.date_of_birth
+      ? new Date(initialData?.user.date_of_birth).toISOString().split("T")[0]
+      : "",
+    learningGoal: initialData?.user.learning_goal || "",
+    businessName: initialData?.user.business_name || "",
+    industry: initialData?.user.industry_id || "",
+    entrepreneurStage: initialData?.user.entrepreneur_stage_id || "",
   });
+
+  // --- Iterate initial data (so it doesn't get lost)
+  useEffect(() => {
+    if (initialData) {
+      setFormData({
+        fullName: initialData.user.full_name || "",
+        email: initialData.user.email || "",
+        avatar: initialData.user.avatar || "",
+        roleId: initialData.user.role.id ?? "",
+        status: initialData.user.status || "ACTIVE",
+        dateOfBirth: initialData?.user.date_of_birth
+          ? new Date(initialData?.user.date_of_birth)
+              .toISOString()
+              .split("T")[0]
+          : "",
+        learningGoal: initialData.user.learning_goal || "",
+        businessName: initialData.user.business_name || "",
+        industry: initialData.user.industry_id || "",
+        entrepreneurStage: initialData.user.entrepreneur_stage_id || "",
+      });
+    }
+  }, [initialData]);
 
   // --- Add event listener to prevent page refresh
   useEffect(() => {
@@ -101,8 +134,13 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
   }, []);
 
   // --- Extract variable
-  const isLoading = isLoadingRoles || isLoadingIndustries || isLoadingStages;
-  const isError = isErrorRoles || isErrorIndustries || isErrorStages;
+  const isLoading =
+    isLoadingInitial ||
+    isLoadingRoles ||
+    isLoadingIndustries ||
+    isLoadingStages;
+  const isError =
+    isErrorInitial || isErrorRoles || isErrorIndustries || isErrorStages;
   if (isLoading) {
     return (
       <div className="flex w-full h-full items-center justify-center text-alternative">
@@ -162,11 +200,12 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
 
     // -- POST to Database
     try {
-      createUser.mutate(
+      editUser.mutate(
         {
+          id: userId,
           full_name: formData.fullName,
+          avatar: formData.avatar.trim() ? formData.avatar : undefined,
           email: formData.email,
-          avatar: formData.avatar?.trim() ? formData.avatar : undefined,
           role_id: Number(formData.roleId),
           status: formData.status,
           date_of_birth: formData.dateOfBirth.trim()
@@ -187,23 +226,13 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
         },
         {
           onSuccess: () => {
-            toast.success("New user created");
-            setFormData({
-              fullName: "",
-              email: "",
-              avatar: "",
-              roleId: "",
-              status: "ACTIVE",
-              dateOfBirth: "",
-              learningGoal: "",
-              businessName: "",
-              industry: "",
-              entrepreneurStage: "",
-            });
-            router.push("/users");
+            toast.success("Edited Succesfully");
+            utils.read.user.invalidate({ id: userId });
+            utils.list.users.invalidate();
+            router.push(`/users/${userId}`);
           },
           onError: (err) => {
-            toast.error("Failed to create user", {
+            toast.error("Failed to update", {
               description: err.message,
             });
           },
@@ -227,16 +256,20 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
           <ChevronRight className="size-3.5" />
           <AppBreadcrumbItem href="/users">Users</AppBreadcrumbItem>
           <ChevronRight className="size-3.5" />
-          <AppBreadcrumbItem href="/users/create" isCurrentPage>
-            Create
+          <AppBreadcrumbItem href={`/users/${userId}`}>
+            Profile
+          </AppBreadcrumbItem>
+          <ChevronRight className="size-3.5" />
+          <AppBreadcrumbItem href={`/users/${userId}/edit`} isCurrentPage>
+            Edit
           </AppBreadcrumbItem>
         </AppBreadcrumb>
         <div className="page-title-actions flex justify-between items-center">
           {/* --- Page Title */}
           <TitleRevealCMS
-            titlePage={"Add New User"}
+            titlePage={"Edit User"}
             descPage={
-              "Fill out the form to add a new user to the system, including basic profile information."
+              "Update user details, adjust roles or permissions, and manage account status easily."
             }
           />
 
@@ -261,7 +294,7 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
               ) : (
                 <Save className="size-5" />
               )}
-              Save User
+              Save Changes
             </AppButton>
           </div>
         </div>
@@ -274,7 +307,10 @@ export default function CreateUserForm({ sessionToken }: CreateUserFormProps) {
           </h2>
 
           {/* --- Upload Avatar */}
-          <UploadAvatarUserCMS onUpload={handleImageForm} />
+          <UploadAvatarUserCMS
+            onUpload={handleImageForm}
+            value={formData.avatar}
+          />
 
           {/* Personal Information Data */}
           <div className="personal-information-data flex gap-5">
