+++ src/app/components/titles/SectionTitleRestart25.jsx
